% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{create_summary_df}
\alias{create_summary_df}
\title{Summary Dataframe
Returns a dataframe of the epidemic duration, proportion loss, exposure index, and
number of affected farms for every simulation produced by vary_params}
\usage{
create_summary_df(filename, val_vector, param_name)
}
\arguments{
\item{filename}{path to file containing results of vary_params}

\item{val_vector}{range of the varied parameter}

\item{param_name}{name of the varied parameter}
}
\description{
Summary Dataframe
Returns a dataframe of the epidemic duration, proportion loss, exposure index, and
number of affected farms for every simulation produced by vary_params
}
