% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_all_sims}
\alias{get_all_sims}
\title{Returns a dataframe of simulation number, time, and number of birds in the given compartment}
\usage{
get_all_sims(compartment, results)
}
\arguments{
\item{compartment}{character representing which compartment (S, I, R, V, or C) to analyze}

\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of simulation number, time, and number of birds in the given compartment
}
