% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_depth2_spread}
\alias{get_depth2_spread}
\title{Depth Two Spread}
\usage{
get_depth2_spread(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of simulation number and whether there are infections along a depth two section of the network
}
