% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_duration_array}
\alias{get_duration_array}
\title{Returns a dataframe of simulation number and epidemic duration, the number of days with at least one infection present.}
\usage{
get_duration_array(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of simulation number and epidemic duration, the number of days with at least one infection present.
}
