% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_exposure_fraction}
\alias{get_exposure_fraction}
\title{Fraction Infected}
\usage{
get_exposure_fraction(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of the simulation number and the fraction of birds infected
}
