% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_failure_array}
\alias{get_failure_array}
\title{Returns a dataframe of simulation number and whether the outbreak failed (defined as at most 1 infection)}
\usage{
get_failure_array(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of simulation number and whether the outbreak failed (defined as at most 1 infection)
}
