% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_farm_culls}
\alias{get_farm_culls}
\title{Number of Infected Farms}
\usage{
get_farm_culls(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a vector of the number of farms culled in each simulation
}
