% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_infectious_array}
\alias{get_infectious_array}
\title{Infected Statistics}
\usage{
get_infectious_array(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of simulation number and lower/median/upper number of infected chickens
}
