% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_n_spread}
\alias{get_n_spread}
\title{Retrieve N Spread of Infections}
\usage{
get_n_spread(results, spread_number)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}

\item{spread_number}{number of nodes spread to examine}
}
\description{
Returns a dataframe of whether or not infections have spread to N nodes in the network.
}
