% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_proportion_loss}
\alias{get_proportion_loss}
\title{Returns a dataframe of the simulation number and the proportion of chickens lost to disease or culling}
\usage{
get_proportion_loss(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of the simulation number and the proportion of chickens lost to disease or culling
}
