% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_susceptibles_array}
\alias{get_susceptibles_array}
\title{Returns a dataframe of simulation number and lower/median/upper number of susceptible chickens}
\usage{
get_susceptibles_array(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of simulation number and lower/median/upper number of susceptible chickens
}
