% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{make_graphs}
\alias{make_graphs}
\title{Summarizing Graphs}
\usage{
make_graphs(results_list, param_values)
}
\arguments{
\item{results_list}{list of results}

\item{param_values}{list of independent variable}
}
\description{
Summarizes the results of vary_params by graphing bird loss, farms affected, outbreak duration,
and number of birds exposed against the given range of values.
}
