% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaflu-package.R
\docType{package}
\name{metaflu}
\alias{metaflu}
\alias{metaflu-package}
\title{metaflu.- Stochastic simulator for avian influenza metapopulation dynaimics}
\description{
metaflu.- Stochastic simulator for avian influenza metapopulation dynaimics
}
