% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{mf_sim}
\alias{mf_sim}
\title{Influenza Metapopulation Simulator}
\usage{
mf_sim(init, parameters, times, n_sims = 1, seed = NULL, parallel = FALSE,
  progress = FALSE)
}
\arguments{
\item{init}{Intital conditions - a patches X 4 (SIRV) matrix of populations}

\item{parameters}{a list of parameter values.  See below for details}

\item{times}{a vector of output times to report}

\item{n_sims}{number of simulations to run}

\item{seed}{if set, a random seed for the simulations}

\item{parallel}{run the simulations in parallel?}

\item{progress}{TRUE/FALSE show a progress bar?}
}
\description{
Influenza Metapopulation Simulator
}

