% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{proportion_failed}
\alias{proportion_failed}
\title{Returns the proportion of epidemic failures, define as the number of failures/total results}
\usage{
proportion_failed(failure_results)
}
\arguments{
\item{failure_results}{the results of a get_failure_array() call}
}
\description{
Returns the proportion of epidemic failures, define as the number of failures/total results
}
