% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{reduce_epi_array}
\alias{reduce_epi_array}
\title{Returns array}
\usage{
reduce_epi_array(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns array
}
