% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{vary_params}
\alias{vary_params}
\title{Produces graphs showing how bird loss, farms affected, outbreak duration, and number of birds exposed changes
as the given parameter varies along the given range}
\usage{
vary_params(param_name, param_values, sims, num_of_farms, num_of_chickens,
  parms)
}
\arguments{
\item{sims}{the number of simulations to run for each value of the parameter}

\item{parms}{the list of parameters}

\item{param_value}{the name of the parameter to vary}

\item{param_vector}{the range over which the parameter will vary}

\item{farm_num}{the number of farms in the network}

\item{farm_size}{the typical farm size}
}
\description{
Produces graphs showing how bird loss, farms affected, outbreak duration, and number of birds exposed changes
as the given parameter varies along the given range
}
