% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{vary_params_2d}
\alias{vary_params_2d}
\title{2-Dimensional Parameter Variation}
\usage{
vary_params_2d(param1_name, param1_values, param2_name, param2_values,
  sims = 1000, num_of_farms = 200, num_of_chickens = 50, parms)
}
\arguments{
\item{sims}{the number of simulations to run for each value of the parameter}

\item{parms}{the list of parameters}

\item{param_value}{the name of the parameter to vary}

\item{param_vector}{the range over which the parameter will vary}

\item{farm_num}{the number of farms in the network}

\item{farm_size}{the typical farm size}
}
\description{
Returns an array of the results of varying the 2 given parameters over the given ranges.
}
