% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizing_functions.R
\name{get_number_farms}
\alias{get_number_farms}
\title{Number of Affected Farms}
\usage{
get_number_farms(results)
}
\arguments{
\item{results}{4-dimensional array of results (compartment, patch, time, simulation)}
}
\description{
Returns a dataframe of the simulation number and the number of farms affected by the outbreak
}
