% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{mf_sim}
\alias{mf_sim}
\title{Influenza Metapopulation Simulator}
\usage{
mf_sim(parameters, network, times = NULL, n_sims = 1, seed = NULL,
  .parallel = FALSE)
}
\description{
Influenza Metapopulation Simulator
}

